refactor(session): merge bus DB accessors, rename record-fields
- bus-persistence/query-database.ts folded into database.ts: both
  accessors returned the same telemetry-first handle; query-side
  importers repointed (4 files)
- bus-persistence/record-helpers.ts -> record-fields.ts (4 importers)
- ledger-core/db.ts relocated to test/ledger-core/db.ts: verified the
  ledger DDL drift gate consumes schema.ts directly, and the only
  consumer of createLedgerDb is append.test.ts; drizzle.config.ts
  comment repointed
