+++ packages/session/test/ledger-core/db.ts
@@ -1,6 +1,6 @@
 import type { Database } from "bun:sqlite";
 import { drizzle } from "drizzle-orm/bun-sqlite";
-import { ledgerEvent, ledgerHead } from "./schema";
+import { ledgerEvent, ledgerHead } from "../../src/ledger-core/schema";
 
 const schema = { ledgerEvent, ledgerHead } as const;
 
@@ -14,6 +14,10 @@ const schema = { ledgerEvent, ledgerHead } as const;
  * Phase A scope: drizzle is DDL/schema-parity surface only. Decision-class
  * runtime SQL (CAS, changes===1 receipt, hash chain) lives in append.ts /
  * chain.ts as raw prepared statements and must stay there.
+ *
+ * Test-only fixture: no production code consumes this factory, and the
+ * ledger DDL drift gate (script/check-ledger-schema-drift.ts) consumes
+ * schema.ts directly, not this client.
  */
 export function createLedgerDb(db: Database) {
   return drizzle({ client: db, schema, casing: "snake_case" });
